iio: imu: Add driver for Bosch BMI260 IMU
The ultra-low power BMI260 is an IMU consisting of a 16-bit tri-axial
gyroscope and a 16-bit tri-axial accelerometer combining precise
acceleration, angular rate measurement and intelligent on-chip
motion-triggered interrupt features.

The driver supports the BMI260 over I2C. It is based on the BMI160
driver, and like that driver supports accelerometer and gyroscope
reading, as well as data ready interrupts.