PM: hibernate: Enter s2idle sleep state before hibernation
Windows enters the s2idle screen_off state before hibernation, which
turns off e.g., the keyboard light and in some devices pulses the
suspend light. It may also turn off auxiliaries prior to hibernation.
These auxiliaries are 1) not used to maintain the swap of the system
and 2) can misbehave during the wake-ups in hibernation.

One example is the OneXPlayer X1 controller, which will have its
vibration motor stuck on, if it is not first deactivated by entering
the sleep state.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>